All Assignments from SENG 265
a1 & a2 are in C and build off each other. Getting data from a survey on a text file and sorting and printing it all out to it is easy to go over.

a3, a4, & a5 are coded in Python and build off each other. Making a database to store patients files with notes.